+++ client/src/components/forms/utils/index.js
@@ -13,3 +13,4 @@ export { DynamicRender } from "./DynamicRender";
 export { SecondaryButton } from "./SecondaryButton";
 export { default as StatefulTabs } from "./StatefulTabs";
 export { ToolTipWrapper } from "./ToolTip";
+export { CannotBeUndoneModal } from "./CannotBeUndoneModal";