+++ unit-tests/json-to-iac/security-groups.test.js
@@ -121,6 +121,240 @@ resource "ibm_is_security_group" "management_vpc_management_vpe_sg_sg" {
 `;
       assert.deepEqual(actualData, expectedData, "it should return correct tf");
     });
+    it("should format a security group", () => {
+      let actualData = formatSecurityGroup(
+        {
+          vpc: "management",
+          name: "management-vpe-sg",
+          resource_group: "slz-management-rg",
+          rules: [
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "inbound",
+              name: "allow-ibm-inbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "inbound",
+              name: "allow-vpc-inbound",
+              source: "10.0.0.0/8",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-vpc-outbound",
+              source: "10.0.0.0/8",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-53-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 53,
+                port_min: 53,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-80-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 80,
+                port_min: 80,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-443-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 443,
+                port_min: 443,
+              },
+            },
+          ],
+        },
+        {
+          _options: {
+            region: "us-south",
+            tags: ["hello", "world"],
+            prefix: "iac",
+          },
+          resource_groups: [
+            {
+              use_prefix: false,
+              name: "slz-service-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-management-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-workload-rg",
+              use_data: false,
+            },
+          ],
+          vpcs: [
+            {
+              name: "management",
+              use_data: false,
+            },
+          ],
+        }
+      );
+      let expectedData = `
+resource "ibm_is_security_group" "management_vpc_management_vpe_sg_sg" {
+  name           = "\${var.prefix}-management-management-vpe-sg-sg"
+  vpc            = ibm_is_vpc.management_vpc.id
+  resource_group = var.slz_management_rg_id
+  tags = [
+    "hello",
+    "world"
+  ]
+}
+`;
+      assert.deepEqual(actualData, expectedData, "it should return correct tf");
+    });
+    it("should format a security group from data", () => {
+      let actualData = formatSecurityGroup(
+        {
+          vpc: "management",
+          name: "management-vpe-sg",
+          resource_group: "slz-management-rg",
+          use_data: true,
+          rules: [
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "inbound",
+              name: "allow-ibm-inbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "inbound",
+              name: "allow-vpc-inbound",
+              source: "10.0.0.0/8",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-vpc-outbound",
+              source: "10.0.0.0/8",
+              tcp: {
+                port_max: null,
+                port_min: null,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-53-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 53,
+                port_min: 53,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-80-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 80,
+                port_min: 80,
+              },
+            },
+            {
+              vpc: "management",
+              sg: "management-vpe-sg",
+              direction: "outbound",
+              name: "allow-ibm-tcp-443-outbound",
+              source: "161.26.0.0/16",
+              tcp: {
+                port_max: 443,
+                port_min: 443,
+              },
+            },
+          ],
+        },
+        {
+          _options: {
+            region: "us-south",
+            tags: ["hello", "world"],
+            prefix: "iac",
+          },
+          resource_groups: [
+            {
+              use_prefix: false,
+              name: "slz-service-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-management-rg",
+              use_data: false,
+            },
+            {
+              use_prefix: false,
+              name: "slz-workload-rg",
+              use_data: false,
+            },
+          ],
+          vpcs: [
+            {
+              name: "management",
+              use_data: true,
+            },
+          ],
+        }
+      );
+      let expectedData = `
+data "ibm_is_security_group" "management_vpc_management_vpe_sg_sg" {
+  name = "management-vpe-sg"
+  vpc  = data.ibm_is_vpc.management_vpc.id
+}
+`;
+      assert.deepEqual(actualData, expectedData, "it should return correct tf");
+    });
   });
   describe("formatSgRule", () => {
     it("should format security group rule with no protocol", () => {
@@ -167,6 +401,66 @@ resource "ibm_is_security_group" "management_vpc_management_vpe_sg_sg" {
       let expectedData = `
 resource "ibm_is_security_group_rule" "management_vpc_management_vpe_sg_sg_rule_allow_ibm_inbound" {
   group     = ibm_is_security_group.management_vpc_management_vpe_sg_sg.id
+  remote    = "161.26.0.0/16"
+  direction = "inbound"
+}
+`;
+      assert.deepEqual(
+        actualData,
+        expectedData,
+        "it should return correct data"
+      );
+    });
+    it("should format security group rule with no protocol and security group uses", () => {
+      let actualData = formatSgRule(
+        {
+          sg: "management-vpe-sg",
+          vpc: "management",
+          action: "allow",
+          destination: "10.0.0.0/8",
+          direction: "inbound",
+          name: "allow-ibm-inbound",
+          source: "161.26.0.0/16",
+          icmp: {
+            type: null,
+            code: null,
+          },
+          tcp: {
+            port_min: null,
+            port_max: null,
+            source_port_min: null,
+            source_port_max: null,
+          },
+          udp: {
+            port_min: null,
+            port_max: null,
+            source_port_min: null,
+            source_port_max: null,
+          },
+        },
+        {
+          _options: {
+            region: "us-south",
+            prefix: "iac",
+            tags: ["hello", "world"],
+          },
+          resource_groups: [
+            {
+              use_data: false,
+              name: "slz-management-rg",
+            },
+          ],
+          security_groups: [
+            {
+              name: "management-vpe-sg",
+              use_data: true,
+            },
+          ],
+        }
+      );
+      let expectedData = `
+resource "ibm_is_security_group_rule" "management_vpc_management_vpe_sg_sg_rule_allow_ibm_inbound" {
+  group     = data.ibm_is_security_group.management_vpc_management_vpe_sg_sg.id
   remote    = "161.26.0.0/16"
   direction = "inbound"
 }